+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Features
 - defaults for ``date`` and software base attributes #657
 - ``Series::setSoftware()`` add second argument for version #657
 - free standing functions to query the API version and feature variants at runtime #665
+- expose ``determineFormat`` and ``suffix`` functions #684
 
 Bug Fixes
 """""""""